Create interface.ts and firebaseFunctions.ts
